feature/add source command
This command links to the code of a command if it can be found. The embed could use a little work as it returns an empty description, this description shoul dbe the  of a command. There is a known issue with the embed where entering an invalid cog name raises an exception, that exception escapes the markdown in the embed, it is mildly annoying.